pyproject.toml
[project]
name = "merit-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mark Reith", email = "mark@appmerit.com" },
    { name = "Nikita Shirobokov", email = "nick@appmerit.com" }
]
requires-python = ">=3.12"
dependencies = [
    "openai>=2.6.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "hdbscan>=0.8.38",
    "anthropic[vertex]>=0.72.0",
    "openai-agents>=0.4.2",
    "claude-agent-sdk>=0.1.6",
    "boto3>=1.40.67",
    "botocore>=1.40.67",
    "sentence-transformers>=5.1.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
lint = [
    "mypy==1.18.2",
    "ruff==0.14.5",
]

[project.scripts]
merit-analyzer = "merit_analyzer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["ALL",]
ignore = [
    "COM812", # conflicts with fomatters
    "RUF012", # Doesn't play well with Pydantic
    "TC001",  # Doesn't play well with Pydantic
    "TC002",  # Doesn't play well with Pydantic
    "TC003",  # Doesn't play well with Pydantic
    "D107",   # Missing docstring in __init__
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "D",
    "S", 
    "PLR2004",
    "ANN", # TODO: Relax type annotations in tests
    "T201", # Allow print statements in tests
    "INP001", # __init__ not required in tests
]

[tool.ruff.format]
quote-style = "double" # Like Black, use double quotes for strings
indent-style = "space" # Like Black, use spaces for indentation
docstring-code-format = true
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
combine-as-imports = true
lines-after-imports = 2

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = ["pydantic.mypy"]
strict = true
enable_error_code = "deprecated"
